+++ src/Shared/Avalon.Network.Packets/Auth/SWorldHandshakePacket.cs
@@ -10,10 +10,10 @@ public class SWorldHandshakePacket : Packet
     public static NetworkProtocol Protocol = NetworkProtocol.Tcp;
     public static NetworkPacketFlags Flags = NetworkPacketFlags.Encrypted;
     
-    [ProtoMember(1)] public int? AccountId { get; set; }
+    [ProtoMember(1)] public ulong? AccountId { get; set; }
     [ProtoMember(2)] public bool Verified { get; set; }
 
-    public static NetworkPacket Create(int accountId, bool verified, Func<byte[], byte[]> encryptFunc)
+    public static NetworkPacket Create(ulong accountId, bool verified, Func<byte[], byte[]> encryptFunc)
     {
         using var memoryStream = new MemoryStream();
