+++ src/Shared/Avalon.Network.Packets/Map/CMapTeleportPacket.cs
@@ -10,11 +10,11 @@ public class CMapTeleportPacket : Packet
     public static NetworkProtocol Protocol = NetworkProtocol.Tcp;
     public static NetworkPacketFlags Flags = NetworkPacketFlags.Encrypted;
     
-    [ProtoMember(1)] public int AccountId { get; set; }
-    [ProtoMember(2)] public int CharacterId { get; set; }
+    [ProtoMember(1)] public ulong AccountId { get; set; }
+    [ProtoMember(2)] public ulong CharacterId { get; set; }
     [ProtoMember(3)] public int MapId { get; set; }
 
-    public static NetworkPacket Create(int accountId, int characterId, int mapId, Func<byte[], byte[]> encryptFunc)
+    public static NetworkPacket Create(ulong accountId, ulong characterId, int mapId, Func<byte[], byte[]> encryptFunc)
     {
         using var memoryStream = new MemoryStream();
